+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - server: support database-to-database joins with BigQuery
 - server: improved startup time when using large remote schemas
 - server: fix rest-endpoints bug allow list arguments (fix #7135)
+- server: fallback to unauthorized role when JWT is not found in cookie (fix #7272)
 - console: add comments to tracked functions
 - console: add select all columns option while selecting the columns in event triggers
 - console: add request transforms for events